haciendo que el script se ejecute suponiendo que el servidor está encendido

+++ validar-echo-server.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
 CONF_FILE="server/config.ini"
-BACKUP_FILE="config.ini.bak"
 PORT="12345"
 
 if [ ! -f "$CONF_FILE" ]; then
@@ -9,12 +8,16 @@ if [ ! -f "$CONF_FILE" ]; then
     exit 1
 fi
 
-cp "$CONF_FILE" "$BACKUP_FILE"
-echo -e "[DEFAULT]\nSERVER_PORT = $PORT\nSERVER_IP = server\nSERVER_LISTEN_BACKLOG = 5\nLOGGING_LEVEL = INFO" > "$CONF_FILE"
+source "$CONF_FILE"
+
+if [ -z "$PORT" ]; then
+  echo "Error: el archivo $CONF_FILE debe definir PORT"
+  exit 1
+fi
+
 
 MSG="test message"
 
-sudo docker compose -f docker-compose-dev.yaml up server -d
 OUTPUT=$(echo "$MSG" | sudo docker run -i --rm --network=tp0_testing_net busybox nc server "$PORT")
 
 if [ "$OUTPUT" = "$MSG" ]; then
@@ -22,7 +25,3 @@ if [ "$OUTPUT" = "$MSG" ]; then
 else
     echo "action: test_echo_server | result: fail"
 fi
-
-sudo docker compose -f docker-compose-dev.yaml down --remove-orphans
-mv -f "$BACKUP_FILE" "$CONF_FILE"
-
